Revised the constraint controller. Add rotation constraint for a joint. Can also specify a local offset ( either position or rotation ) for the constraint. ex : bml char doctor <sbm:constraint sbm:handle="cdoctor" effector="l_wrist" sbm:constraint-type="rot" target="foo" rot-x="180" rot-z="180" rot-y="10"/>
Revised the position control UI in the debug viewer. Now the user can drag  along each axis to move the object along that axis.
Add rotation control UI to change the orientation for a pawn object. (Shift + LMB)

Revised example-based reach controller ( sbm:bodyreach ). Add procedural IK animation when no motion data is given. Add initial bml command to use the controller. 
ex : bml char doctor <sbm:bodyreach sbm:handle="creach" target="foo"/>

+++ core/smartbody/smartbody-lib/src/sbm/me_ct_data_driven_reach.hpp
@@ -96,7 +96,7 @@ template <class T, class S> void pairToVec(const vector<pair<T,S>>& A, vector<T>
 {
 	outT.resize(A.size());
 	outS.resize(A.size());
-	for (size_t i=0;i<A.size();i++)
+	for (unsigned int i=0;i<A.size();i++)
 	{
 		outT[i] = A[i].first;
 		outS[i] = A[i].second;
@@ -106,7 +106,7 @@ template <class T, class S> void vecToPair(vector<T>& vecT, vector<S>& vecS, vec
 {
 	int numElem = vecT.size() < vecS.size() ? vecT.size() : vecS.size();
 	outVec.resize(numElem);
-	for (size_t i=0;i<outVec.size();i++)
+	for (unsigned int i=0;i<outVec.size();i++)
 	{
 		outVec[i].first = vecT[i];
 		outVec[i].second = vecS[i];
